+++ src/app/core/auth/interfaces/customer-response.ts
@@ -1,12 +1,9 @@
-import Address from '../model/address';
+import { Address } from '../../model/address';
 
 export interface CustomerResponse {
   customer: {
-    email: string;
-    password: string;
     firstName: string;
     lastName: string;
-    dateOfBirth: string;
     addresses: Address[];
   };
 }
