+++ src/app/core/model/address.ts
@@ -1,8 +1,8 @@
-export default interface Address {
+export interface Address {
   streetName: string;
   city: string;
   postalCode: string;
   country: string;
   isDefaultBillingAddress?: boolean;
   isDefaultShippingAddress?: boolean;
-} // eslint-disable-line semi
+}
