+++ src/app/core/model/user.ts
@@ -1,10 +1,10 @@
-import Address from './address';
+import { Address } from './address';
 
-export default interface User {
+export interface User {
   email: string;
   password: string;
   firstName?: string;
   lastName?: string;
   dateOfBirth?: string;
   addresses?: Address[];
-} // eslint-disable-line semi
+}
